Try moving sidebar navigation context up higher
